Removing the JDBC Connector EJB and refactoring all the code to org.silverpeas
