--- jdbc-connector/jdbc-connector-ejb/src/main/java/com/stratelia/silverpeas/connecteurJDBC/control/ConnecteurJDBCException.java
@@ -1,68 +1,0 @@
-/**
- * Copyright (C) 2000 - 2011 Silverpeas
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * As a special exception to the terms and conditions of version 3.0 of
- * the GPL, you may redistribute this Program in connection with Free/Libre
- * Open Source Software ("FLOSS") applications as described in Silverpeas's
- * FLOSS exception.  You should have received a copy of the text describing
- * the FLOSS exception, and it is also available here:
- * "http://repository.silverpeas.com/legal/licensing"
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package com.stratelia.silverpeas.connecteurJDBC.control;
-
-import com.stratelia.webactiv.util.exception.SilverpeasException;
-
-/**
- * Title: Connecteur JDBC Description: Ce composant a pour objet de permettre de recuperer
- * rapidement et simplement des donnees du systeme d'information de l'entreprise.
- * @author Eric BURGEL
- * @version 1.0
- * @modified by Mohammed Hguig
- */
-
-public class ConnecteurJDBCException extends SilverpeasException {
-
-  /**
-   * 
-   */
-  private static final long serialVersionUID = 1L;
-
-  public ConnecteurJDBCException(String callingClass, int errorLevel,
-      String message) {
-    super(callingClass, errorLevel, message);
-  }
-
-  public ConnecteurJDBCException(String callingClass, int errorLevel,
-      String message, String extraParams) {
-    super(callingClass, errorLevel, message, extraParams);
-  }
-
-  public ConnecteurJDBCException(String callingClass, int errorLevel,
-      String message, Exception nested) {
-    super(callingClass, errorLevel, message, nested);
-  }
-
-  public ConnecteurJDBCException(String callingClass, int errorLevel,
-      String message, String extraParams, Exception nested) {
-    super(callingClass, errorLevel, message, extraParams, nested);
-  }
-
-  public String getModule() {
-    return "connecteurJDBC";
-  }
-
-}
