--- jdbc-connector/jdbc-connector-ejb/src/main/resources/META-INF/ejb-jar.xml
@@ -1,59 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (C) 2000 - 2011 Silverpeas
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU Affero General Public License as
-    published by the Free Software Foundation, either version 3 of the
-    License, or (at your option) any later version.
-
-    As a special exception to the terms and conditions of version 3.0 of
-    the GPL, you may redistribute this Program in connection with Free/Libre
-    Open Source Software ("FLOSS") applications as described in Silverpeas's
-    FLOSS exception.  You should have received a copy of the text describing
-    the FLOSS exception, and it is also available here:
-    "http://repository.silverpeas.com/legal/licensing"
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU Affero General Public License for more details.
-
-    You should have received a copy of the GNU Affero General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
--->
-<ejb-jar id="ejb-jar_ID" version="2.1" xmlns="http://java.sun.com/xml/ns/j2ee"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" 
-         xsi:schemaLocation="http://java.sun.com/xml/ns/j2ee http://java.sun.com/xml/ns/j2ee/ejb-jar_2_1.xsd">
-  <description>no description</description>
-  <display-name>ConnecteurJDBCJar</display-name>
-  <enterprise-beans>
-    <session>
-      <description>no description</description>
-      <display-name>ConnecteurJDBCBm</display-name>
-      <ejb-name>ConnecteurJDBCBm</ejb-name>
-      <home>com.stratelia.silverpeas.connecteurJDBC.control.ejb.ConnecteurJDBCBmHome</home>
-      <remote>com.stratelia.silverpeas.connecteurJDBC.control.ejb.ConnecteurJDBCBm</remote>
-      <ejb-class>com.stratelia.silverpeas.connecteurJDBC.control.ejb.ConnecteurJDBCBmEJB</ejb-class>
-      <session-type>Stateless</session-type>
-      <transaction-type>Container</transaction-type>
-      <resource-ref>
-        <res-ref-name>jdbc/Silverpeas</res-ref-name>
-        <res-type>javax.sql.DataSource</res-type>
-        <res-auth>Container</res-auth>
-      </resource-ref>
-    </session>
-  </enterprise-beans>
-  <assembly-descriptor>
-    <container-transaction>
-      <method>
-        <ejb-name>ConnecteurJDBCBm</ejb-name>
-        <method-name>*</method-name>
-      </method>
-      <trans-attribute>NotSupported</trans-attribute>
-    </container-transaction>
-  </assembly-descriptor>
-</ejb-jar>
-
