+++ jdbc-connector/jdbc-connector-jar/pom.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>com.silverpeas.components</groupId>
+    <artifactId>jdbc-connector</artifactId>
+    <version>5.9-SNAPSHOT</version>
+  </parent>
+
+  <groupId>com.silverpeas.components.jdbc-connector</groupId>
+  <artifactId>jdbc-connector-jar</artifactId>
+  <name>Jdbc-connector Jar</name>
+
+  <dependencies>
+    <dependency>
+      <groupId>com.silverpeas.core</groupId>
+      <artifactId>lib-core</artifactId>
+      <version>${core.version}</version>
+    </dependency>
+  </dependencies>
+</project>
