+++ jdbc-connector/jdbc-connector-war/src/main/webapp/connecteurJDBC/jsp/imports.jsp
@@ -34,12 +34,11 @@
 <%@ page import="com.stratelia.webactiv.util.viewGenerator.html.frame.Frame"%>
 <%@ page import="com.stratelia.webactiv.util.viewGenerator.html.icons.*"%>
 <%@ page import="com.stratelia.webactiv.util.viewGenerator.html.iconPanes.*"%>
-<%@ page import="com.stratelia.webactiv.util.viewGenerator.html.arrayPanes.*"%>
 <%@ page import="com.stratelia.webactiv.util.viewGenerator.html.browseBars.*"%>
 <%@ page import="com.stratelia.webactiv.util.viewGenerator.html.operationPanes.*"%>
 <%@ page import="com.stratelia.webactiv.util.viewGenerator.html.buttonPanes.ButtonPane"%>
 <%@ page import="com.stratelia.webactiv.util.viewGenerator.html.operationPanes.OperationPane"%>
-<%@ page import="com.stratelia.silverpeas.connecteurJDBC.control.ConnecteurJDBCSessionController"%>
+<%@ page import="org.silverpeas.connecteurJDBC.control.ConnecteurJDBCSessionController"%>
 <%@ page import="com.stratelia.webactiv.util.GeneralPropertiesManager"%>
 <%@ page import="com.stratelia.webactiv.util.*"%>
 <%@ page import="com.stratelia.silverpeas.silvertrace.*"%>
