+++ jdbc-connector/jdbc-connector-war/src/test/resources/spring-jdbc-connector.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    Copyright (C) 2000 - 2011 Silverpeas
+    Copyright (C) 2000 - 2009 Silverpeas
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU Affero General Public License as
@@ -25,16 +25,22 @@
 
 -->
 
-<web-app>
-  <display-name>ConnecteurJDBC War</display-name>
-  <servlet>
-    <display-name>ConnecteurJDBCRequestRouter</display-name>
-    <servlet-name>ConnecteurJDBCRequestRouter</servlet-name>
-    <servlet-class>org.silverpeas.connecteurJDBC.control.ConnecteurJDBCRequestRouter</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>ConnecteurJDBCRequestRouter</servlet-name>
-    <url-pattern>/RconnecteurJDBC/*</url-pattern>
-  </servlet-mapping>
-</web-app>
+<beans xmlns="http://www.springframework.org/schema/beans"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xmlns:p="http://www.springframework.org/schema/p"
+       xmlns:ctx="http://www.springframework.org/schema/context"
+       xsi:schemaLocation="
+        http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
+        http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
+
+  <ctx:annotation-config/>
+  <ctx:component-scan base-package="org.silverpeas.connecteurJDBC.control"/>
+  <ctx:component-scan base-package="org.silverpeas.connecteurJDBC.mock"/>
+
+  <!-- the ClassifiedService factory from which Silverpeas components access the classifieds service.
+       It delegate to the Spring container the management of the life-cycle of the ClassifiedService objects. -->
+  <bean id="connecteurJDBCServiceFactory" class="org.silverpeas.connecteurJDBC.control.ConnecteurJDBCServiceFactory"
+        factory-method="getFactory" />
+
+</beans>
